[pb62-frontend] Move useBattle and useBattleAction hooks into battleArena data folder
